blogs can only be deleted by the user who added it

+++ part4/4.1bloglist/controllers/blogs.js
@@ -43,15 +43,30 @@ blogsRouter.post('/', async (req, res) => {
 })
 
 blogsRouter.delete('/:id', async (req, res) => {
+  const token = jwt.verify(req.token, process.env.SECRET)
+  if(!token.id) {
+    return res.status(401).json({ error: 'token invalid' })
+  }
+
+  const user = await User.findById(token.id)
+  if (!user) {
+    return res.status(400).json({ error: 'user ID is missing or not valid'})
+  }
+
   try {
-    const result = await Blog.findByIdAndDelete(req.params.id)
-    if (!result) {
+    const blogToDelete = await Blog.findById(req.params.id)
+    if (!blogToDelete) {
       res.status(404).end() // blog doesn't exist
     }
+
+    if (user.id !== blogToDelete.user.toString()) {
+      res.status(401).json({ error: 'user id that sent delete request is not owner of blog' })
+    }
+    const result = await Blog.findByIdAndDelete(req.params.id)
     res.status(204).end() // successful delete
 
   } catch (err) {
-    res.status(400).end() // blog id is malformatted
+    res.status(400).json({ error: err }) // blog id is malformatted
   }
 })
